+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## master (unreleased)
 
 * Added a "language count" on the homepage (#43).
-* Build a raw, but readable PDF for every language (#21).
+* Build a raw, but readable PDF for every language (#21, #46).
 
 ## 2.3.0 (2016-09-16)